pyproject.toml
[project]
name = "theo"
version = "0.1.0"
description = "Personal AI agent"
readme = "README.md"
authors = [{ name = "Filipe Lima", email = "me@fpl0.io" }]
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.49",
    "asyncpg>=0.31",
    "pgvector>=0.4.2",
    "mlx>=0.31",
    "numpy>=2.4",
    "huggingface-hub>=1.8",
    "tokenizers>=0.22",
    "pydantic>=2.12",
    "pydantic-settings>=2.13",
    "python-dotenv>=1",
    "opentelemetry-api>=1.40",
    "opentelemetry-sdk>=1.40",
    "opentelemetry-exporter-otlp-proto-http>=1.40",
    "opentelemetry-instrumentation-asyncpg>=0.61b0",
    "mlx-whisper>=0.4",
    "aiogram>=3.20",
]

[dependency-groups]
dev = [
    "ruff>=0.15",
    "ty>=0.0.24",
    "sqlfluff>=3",
    "pytest>=9",
    "pytest-asyncio>=0.24",
]

[project.scripts]
theo = "theo.__main__:main"

[build-system]
requires = ["uv_build>=0.11.1,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py314"
line-length = 99

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",       # pydocstyle — we'll add docstring rules later if needed
    "ANN",     # flake8-annotations — ty handles typing
    "COM812",  # trailing comma — conflicts with ruff formatter
    "ISC001",  # single-line implicit string concat — conflicts with formatter
    "T20",     # flake8-print — we use print/logging intentionally
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106", "S108", "PLC0415", "PLR2004", "SLF001"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic_settings.BaseSettings"]

[tool.ty]
rules = { all = "error" }

[tool.ty.environment]
python-version = "3.14"
python-platform = "darwin"

# pydantic-settings injects field values from environment variables at runtime;
# ty cannot model this and flags the constructor as missing required args.
[[tool.ty.overrides]]
include = ["src/theo/config.py"]
rules = { missing-argument = "ignore" }

# Anthropic SDK uses discriminated unions (event.type == "text") that ty
# cannot narrow — it sees the full union and flags attribute access.
[[tool.ty.overrides]]
include = ["src/theo/llm.py"]
rules = { unresolved-attribute = "ignore" }

# pydantic-settings constructors accept str for SecretStr and dynamic kwargs
# like _env_file that ty cannot model.
[[tool.ty.overrides]]
include = ["tests/test_config.py", "tests/test_llm.py", "tests/test_main.py", "tests/test_onboarding.py"]
rules = { invalid-argument-type = "ignore", missing-argument = "ignore", unknown-argument = "ignore" }

# test_retrieval also needs unknown-argument for _env_file kwarg.
[[tool.ty.overrides]]
include = ["tests/test_retrieval.py"]
rules = { invalid-argument-type = "ignore", missing-argument = "ignore", unknown-argument = "ignore" }

# test_contradictions intentionally assigns to frozen dataclass to verify it raises.
[[tool.ty.overrides]]
include = ["tests/test_contradictions.py"]
rules = { invalid-assignment = "ignore" }

# test_edges intentionally assigns to frozen dataclass to verify it raises.
[[tool.ty.overrides]]
include = ["tests/test_edges.py"]
rules = { invalid-assignment = "ignore" }

# test_self_model intentionally assigns to frozen dataclass to verify it raises.
[[tool.ty.overrides]]
include = ["tests/test_self_model.py"]
rules = { invalid-assignment = "ignore" }

# test_episodes intentionally assigns to frozen dataclass to verify it raises.
[[tool.ty.overrides]]
include = ["tests/test_episodes.py"]
rules = { invalid-assignment = "ignore" }

# test_memory_tools uses mock.await_args which ty types as potentially None.
[[tool.ty.overrides]]
include = ["tests/test_memory_tools.py"]
rules = { not-iterable = "ignore" }

# Dict lookups return str, not Literal — ty cannot narrow str keys/values
# from dict[str, ...] to their Literal subtypes.
[[tool.ty.overrides]]
include = ["src/theo/memory/privacy.py"]
rules = { invalid-return-type = "ignore", invalid-assignment = "ignore" }

# pydantic-settings constructors accept str for SecretStr and dynamic kwargs
# like _env_file that ty cannot model.
[[tool.ty.overrides]]
include = ["tests/test_privacy.py"]
rules = { invalid-argument-type = "ignore", invalid-assignment = "ignore" }

# test_telegram uses mock objects and patched constructors that ty cannot model.
[[tool.ty.overrides]]
include = ["tests/test_telegram.py"]
rules = { invalid-argument-type = "ignore", missing-argument = "ignore", unknown-argument = "ignore", invalid-assignment = "ignore", unresolved-attribute = "ignore" }

# Tests that verify frozen dataclasses raise AttributeError on assignment.
# The assignment IS the test — suppression is the only option.
[[tool.ty.overrides]]
include = [
    "tests/test_context.py",
    "tests/test_core_memory.py",
    "tests/test_nodes.py",
    "tests/test_user_model.py",
]
rules = { invalid-assignment = "ignore" }

# test_conversation patches CircuitBreaker.call with an async generator stub.
[[tool.ty.overrides]]
include = ["tests/test_conversation.py"]
rules = { invalid-assignment = "ignore" }

# conftest stubs MLX modules on non-Apple-Silicon CI by setting attributes on
# dynamically created ModuleType instances — ty cannot resolve those attributes.
[[tool.ty.overrides]]
include = ["tests/conftest.py"]
rules = { unresolved-attribute = "ignore" }

[tool.sqlfluff.core]
dialect = "postgres"
max_line_length = 99
exclude_rules = [
    "references.keywords",   # allow unquoted keywords as identifiers (e.g., `label`, `role`)
]

[tool.sqlfluff.indentation]
indent_unit = "space"
tab_space_size = 4

[tool.sqlfluff.layout.type.comma]
line_position = "trailing"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
